+++ src/main/java/com/magvy/experis/javalava_backend/application/DTOs/incoming/MessageDTORequest.java
@@ -1,4 +1,4 @@
-package com.magvy.experis.javalava_backend.application.DTOs;
+package com.magvy.experis.javalava_backend.application.DTOs.incoming;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
