+++ src/main/java/com/magvy/experis/javalava_backend/application/DTOs/outgoing/MessageDTOResponse.java
@@ -1,4 +1,4 @@
-package com.magvy.experis.javalava_backend.application.DTOs;
+package com.magvy.experis.javalava_backend.application.DTOs.outgoing;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.magvy.experis.javalava_backend.domain.entitites.Message;
