+++ src/main/java/com/magvy/experis/javalava_backend/controllers/MessageController.java
@@ -1,6 +1,6 @@
 package com.magvy.experis.javalava_backend.controllers;
-import com.magvy.experis.javalava_backend.application.DTOs.MessageDTORequest;
-import com.magvy.experis.javalava_backend.application.DTOs.MessageDTOResponse;
+import com.magvy.experis.javalava_backend.application.DTOs.incoming.MessageDTORequest;
+import com.magvy.experis.javalava_backend.application.DTOs.outgoing.MessageDTOResponse;
 import com.magvy.experis.javalava_backend.domain.entitites.Message;
 import com.magvy.experis.javalava_backend.domain.entitites.User;
 import com.magvy.experis.javalava_backend.domain.services.MessageService;
